Add tidy up draft creation
